fix renames breaking issues
